requirements.txt
asn1crypto==1.4.0
attrdict==2.0.1
base58==2.0.1
bit==0.7.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coincurve==13.0.0
cryptography==3.3.1
cytoolz==0.11.0
dnspython==2.0.0
ecdsa==0.16.1
email-validator==1.1.2
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
fastapi==0.63.0
hexbytes==0.2.1
huey==2.3.0
idna==2.10
Padding==0.5
parsimonious==0.8.1
pycparser==2.20
pycryptodome==3.9.9
pydantic==1.7.3
pymongo==3.11.2
python-dotenv==0.15.0
redis==3.5.3
requests==2.25.1
rlp==1.2.0
six==1.15.0
starlette==0.13.6
toolz==0.11.1
tronapi==3.1.6
trx-utils==1.0.4
urllib3==1.26.2
